% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fitRGARCH}
\alias{fitRGARCH}
\title{Fits the R-GARCH model to ranking time series data}
\usage{
fitRGARCH(Pi, p = 0, q = 0, inits)
}
\arguments{
\item{Pi}{A matrix of rankings, rows correspond to the time points}

\item{p}{The order of the autoregressive component (default: 0)}

\item{q}{The order of the moving average component (default: 0)}

\item{inits}{Optional initial parameter values for optimization. If not provided, default starting values will be used based on the specified p and q.}
}
\value{
A list containing the simulated rankings, spread trajectory and mean trajectory
}
\description{
Fits the R-GARCH model to ranking time series data
}
