% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lagonedist}
\alias{lagonedist}
\title{Compute the lag-1 Kendall distance for each time point in the ranking time series}
\usage{
lagonedist(Pi)
}
\arguments{
\item{Pi}{A matrix of rankings, rows correspond to the time points}
}
\value{
A vector of lag-1 Kendall distances starting from t=3
}
\description{
Compute the lag-1 Kendall distance for each time point in the ranking time series
}
