% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rRGARCH}
\alias{rRGARCH}
\title{Simulate R-GARCH data}
\usage{
rRGARCH(N, n, phi0, phi = NULL, alpha = NULL, init, burn_in = 100)
}
\arguments{
\item{N}{Length of the time series}

\item{n}{Number of items in the ranking}

\item{phi0}{Intercept term in the mean equation}

\item{phi}{Coefficients for the lagged distance terms in the mean equation}

\item{alpha}{Coefficients for the lagged mean terms in the mean equation}

\item{init}{Initial rankings for the first max(p,q)+1 time points (optional)}

\item{burn_in}{Number of initial time points to discard (default: 100)}
}
\value{
A list containing the simulated rankings, spread trajectory and mean trajectory
}
\description{
Simulate R-GARCH data
}
